Players: create Player interface, rename AbstractPlayer and update children
