+++ game_board/src/fr/lecampusnumerique/cda2025/javaalgo/boardgames/players/Connect4Player.java
@@ -1,10 +1,8 @@
 package fr.lecampusnumerique.cda2025.javaalgo.boardgames.players;
 
 import fr.lecampusnumerique.cda2025.javaalgo.boardgames.symbols.Connect4Symbol;
-import fr.lecampusnumerique.cda2025.javaalgo.boardgames.symbols.Symbols;
-import fr.lecampusnumerique.cda2025.javaalgo.boardgames.symbols.TicTacToeSymbol;
 
-public class Connect4Player extends Player {
+public class Connect4Player extends AbstractPlayer {
 
     public Connect4Player(int number, Connect4Symbol symbol, boolean isArtificial) {
         super(number, symbol, isArtificial);
