+++ game_board/src/fr/lecampusnumerique/cda2025/javaalgo/boardgames/players/TicTacToePlayer.java
@@ -1,9 +1,8 @@
 package fr.lecampusnumerique.cda2025.javaalgo.boardgames.players;
 
-import fr.lecampusnumerique.cda2025.javaalgo.boardgames.symbols.Symbols;
 import fr.lecampusnumerique.cda2025.javaalgo.boardgames.symbols.TicTacToeSymbol;
 
-public class TicTacToePlayer extends Player {
+public class TicTacToePlayer extends AbstractPlayer {
     public TicTacToePlayer(int number, TicTacToeSymbol symbol, boolean isArtificial) {
         super(number, symbol, isArtificial);
     }
